mejoras menores del hito
identacion de codigo en las barras de vida y poco más he movido un cpp que teniamos en los h para evitar problemas.

+++ DeathTokenBase/DeathToken/DialogueBox.h
@@ -15,7 +15,7 @@ class DialogueBox : EventHandler {
 	const int BOXHEIGHT = 250;
 	const int MARGIN = 10;
 	// Definir la altura de línea según la fuente (puede variar)
-	const int lineHeight = Game::FONTSMALLSIZE * 1.3f;  // Se recomienda usar 1.3 * tamaño de la fuente
+	const int lineHeight = static_cast<int>(Game::FONTSMALLSIZE * 1.3f);  // Se recomienda usar 1.3 * tamaño de la fuente
 	// Estimar el ancho promedio de un carácter
 	const float charWidth = Game::FONTSMALLSIZE * 0.6f;  // Aproximadamente 14.4 px para fuente de 24px
 	const int letterdelay = 30; // en milisegundos
